T-2199 Phase D review fix round: C1/C2 criticals, S1-S8 significants, M1-M4/O2, D-SLM3797
Fixes Claude/Poirot/7a3b10a-t2199-phaseD-review.md (FIX-THEN-SHIP, 2 Critical/8 Significant),
merged forward from curie/t2199-phaseD-red@7a3b10a.

C1: sslm_seq_restore never registered in g_live_seqs -- restored sequences read dead forever
(out_tokens=-1, SSLM_OK), greedy included. Fixed: registered at the same point sslm_seq_create
does, before the handle is observable.

C2: alpha_q15 restored to int32_t (plan Sec2.5) and the two-sided [0, 2^20) ceiling
(ValidateDampedGreedyParams) added -- both of the plan's independent overflow defenses now
exist, not one. Propagated through sslm_decode_params, sslm_decode_params_damped_greedy,
DampedGreedyValidationParams, RunGreedyOrDampedGreedyDecodeLoop, and all three struct mirrors
(t2130/t2138 suite headers, Gate C's own identity check).

S1: out_tokens[i]=-1 for a concurrently-dead sequence collided with the documented "pending,
call again" sentinel -- a caller honouring that contract never terminated. Now writes -3, a new,
non-colliding sentinel; sslm_abi_functions_g5_comparable.inc's own contract comment updated to
document three sentinels.

S2: params->mode is now validated -- an unknown mode (neither GREEDY nor DAMPED_GREEDY) rejects
instead of silently degrading to greedy.

S3: D3's own validation-failure path now returns a new SslmForwardStatus::InvalidDecodeParams
(mapped to SSLM_INVALID_ARGUMENT), symmetric with D2's direct return for the identical invalid
set. The "missing Decision-A scale field" rejection is wired at its one real caller:
tools/sslm_generate.cpp's new --decode-mode damped-greedy path reads the artifact's own
DampedGreedyConstants section and rejects cleanly if absent, before deriving q_ln2/q_b/q_c.

S4: tools/sslm_generate.cpp gains --decode-mode greedy|damped-greedy plus
--alpha-q15/--anti-lm-order/--top-k, calling RunGreedyOrDampedGreedyDecodeLoop -- the CLI can
now select damped-greedy mode, closing plan Sec8 D3's stated deliverable. The existing greedy
path is untouched (a separate call site, zero regression risk).

S5: the production surface (DampedGreedyMode, ValidateDampedGreedyParams, the artifact-read
functions, RunGreedyOrDampedGreedyDecodeLoop) relocated from namespace superslm_test_phaseD into
namespace superslm. sslm_decode_step_damped_greedy/sslm_decode_params_damped_greedy are KEPT as
an explicit suite-compatibility shim (phaseD2/phaseD2a/phaseD3's own red-suite cells construct
them directly -- deleting them is red-by-link across three test files, routed to the suite owner
rather than fixed here) -- the real ABI (sslm_decode_step/sslm_decode_params) is now the
documented production-recommended path.

S6: a numeric refusal (SoftmaxKernelRefusedAfterGateAccepted) on a valid model/params no longer
maps to SSLM_ARTIFACT_REJECTED -- a new SSLM_NUMERIC_STEP_REFUSED status, appended at the end of
the closed taxonomy (no renumbering), mirrored into both suite status enums.

S7: the false "not for a call's actual forward-pass work" comment on g_seq_registry_mutex is
corrected. The finding's own suggested reorder (release the registry lock before acquiring
per-sequence locks) was evaluated and found UNSAFE by a concrete counter-example (reintroduces
the exact use-after-free class the D3 fix exists to close) -- documented in place, not
implemented; a real fix needs a different mechanism (e.g. a refcounted/generation-checked
handle), routed as a design question.

S8: build.bat now calls both build_green.bat (Phase A/C) and build_green_phaseD.bat (Phase D) as
real build-time gates (T-2112 probe precedent), via T2199_PHASED_MODEL for the model-needing
cells. build_green_phaseD.bat itself now gates on a nonzero failures= count, not only a crash
(every §4 dispute this ticket routed is resolved, so a failures>0 count is never a fixture
dispute now). t2139_dim9_current_token_pin's own errorlevel gate was already correct; the
remaining piece -- a checkpoint provisioned by DEFAULT so a green battery is impossible with C1
present -- was evaluated (a synthetic-fixture generation pipeline exists,
tools/sslm_pinned_calibration_fixture.py) but not built this round: it would need verifying an
untested fixture round-trips through the decode+save/restore path, and deviates from this
build's own established "assume no real .sslm artifact exists" convention across ~30 other
tools. Routed, not fixed.

M1: sslm_phaseD.h's wrong-constant-name comment corrected during the S5 namespace move.
M2: unresolvable "D-SLM3794A"/"D-SLM3476A" citations corrected to D-SLM3476 (D-SLM3794 dropped --
it rules the artifact flag, not this struct) across sslm_abi.h and both suite mirrors.
M3: free-text damped-greedy's O(vocab_size) mask-narrowing loop skipped when bound_schema is
null (provably a no-op there, all-ones mask) -- the ~19KB allocation itself is kept (still a
required DampedGreedyScoreAndArgmax argument).
M4: documented, not changed -- the mid-loop returns' "state left mutated" is a pre-existing
pattern this same function's every other MapForwardStatus-routed return already has.
O1: not touched -- Dan's call, surfaced not decided.
O2: the plan's own static_cast<int32_t> alpha_eff narrowing added (damped_greedy_topk.cpp) --
value-preserving under C2's now-corrected domain (alpha_eff in [0, 2^20), always int32-safe).

D-SLM3797 (Dan, addendum): sslm_decode_params gains struct_size as its first new field --
caller-set, library-validated (sslm_decode_stepImpl rejects any other value with
SSLM_INVALID_ARGUMENT, checked before every other field). All 15 in-tree production
construction sites (src/damped_greedy_phaseD.cpp, 14 tools/*.cpp smokes/pins) updated to set it
in this same commit; propagated through all three struct mirrors and both Gate C offset-check
lists.

CONFIRMED REGRESSION, ROUTED NOT FIXED: struct_size's strict validation breaks
phaseD2_wiring_red.cpp's own greedy-mode no-regression cross-check (its own `old_params` fixture
predates this addendum and never sets struct_size) -- 12 check failures, confirmed by execution.
The identical code shape (a direct sslm_decode_params construction with no struct_size) exists
across the ENTIRE t2138-abi-red-suite and t2130-g5-red-suite fixture_common.h -- a likely, not
yet suite-executed, consequence for those suites' own decode-step-touching cells. All are test
files outside this build's writable scope; each needs a one-line struct_size addition, which is
the suite owner's own fix, not this build's.

Verified: Phase D suite (real 0.5B checkpoint) -- D1 18/0, D2 54/12 (the struct_size dispute
above, 5 REAL sub-cells otherwise green), D2a 22/0, D3 200/200 trials 1604/0. Phase A/C 278/0/2
unaffected. Full battery (build.bat, no model env vars provisioned) 34189/0 -- the pre-existing
test_main.cpp flags collision this build found last round is already resolved (curie's own fold
23, 7a3b10a, fast-forwarded onto this branch). Never pushed.

+++ src/damped_greedy_topk.cpp
@@ -257,8 +257,16 @@ bool ScoreAndSelect(const int32_t* masked_row, const uint8_t* mask_bits, int32_t
 		const std::size_t si = static_cast<std::size_t>(i);
 		int64_t score;
 		if (MaskBitSet(mask_bits, out->idx[si])) {
-			const int64_t alpha_eff =
-			    (static_cast<int64_t>(alpha_q15) * out->p_omega[si]) >> kProbFracBits;
+			// T-2199 Phase D review fix O2 (Claude/Poirot/7a3b10a-t2199-phaseD-review.md): the
+			// plan's own formula (Sec7.5) narrows alpha_eff through static_cast<int32_t>
+			// before the subtraction -- omitted here previously (kept int64_t throughout).
+			// Value-preserving, not merely spec-matching: C2's fix bounds alpha_q15 to
+			// [0, 2^20) and p_omega to [0, 2^15] by construction (Sec7.2), so alpha_eff's own
+			// true range is [0, 2^20) -- always representable in int32_t, so this cast changes
+			// no computed value under the now-corrected domain, only makes the width the
+			// design specifies explicit in code.
+			const int32_t alpha_eff = static_cast<int32_t>(
+			    (static_cast<int64_t>(alpha_q15) * out->p_omega[si]) >> kProbFracBits);
 			score = out->q_theta[si] - alpha_eff;
 		} else {
 			score = INT64_MIN;
